refactor: extract shared mention autocomplete system from 7 files
The mention/autocomplete logic (typing @ to tag users) was copy-pasted
nearly identically across 7 files — Comment, CommentReplies, Comments,
FeedComment, FeedComments, PostComposer, and [nip19]/+page. Each file
had its own ~300 lines of duplicated functions, state variables, HTML
markup, and CSS for the same feature. Every bug fix required patching
all 7 files independently, and upstream PR #143 demonstrated this
problem by applying the same 3 fixes to each file separately.

This refactor extracts the shared logic into reusable modules:

- src/lib/mentionUtils.ts: pure utility functions (text parsing, DOM
  manipulation, mention rendering)
- src/lib/mentionComposer.ts: MentionComposerController class that
  encapsulates all stateful mention logic per composer instance
- src/components/MentionDropdown.svelte: shared dropdown UI component

Each consumer now instantiates a controller and delegates to it,
reducing ~300 lines of mention code per file to ~25 lines of wiring.

Also wires in the existing but previously unused followListCache.ts
singleton, eliminating redundant kind:3 contact list fetches when
multiple comment boxes are on screen.

Incorporates all 3 fixes from upstream PR #143:
- Regex change to allow spaces in mention queries (e.g. "Zap Cooking")
- Leading space insertion when mention follows non-whitespace text
- Mention dropdown z-index increased from 50 to 1000

Net impact: ~4,650 lines removed across the codebase.
